添加question页面回复模块

+++ src/main/java/life/majiang/community/dto/QuestionDTO.java
@@ -6,15 +6,15 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private long id;
-    private String  title;
-    private String  description;
-    private String  tag;
-    private Long  gmtCreate;
-    private Long  gmtModified;
-    private Integer  creator;
-    private Integer  viewCount;
-    private Integer  commentCount;
-    private Integer  likeCount;
+    private Long id;
+    private String title;
+    private String description;
+    private String tag;
+    private Long gmtCreate;
+    private Long gmtModified;
+    private Long creator;
+    private Integer viewCount;
+    private Integer commentCount;
+    private Integer likeCount;
     private User user;
 }
